+++ Sources/TagKit/Views/TagCapsule.swift
@@ -8,12 +8,10 @@
 
 import SwiftUI
 
-/**
- This view can be used to render tags within a capsule shape.
-
- The view will not slugify the provided tag string, only use
- the content you provide it with.
- */
+/// This view can be used to render tags in a capsule shape.
+///
+/// This view will not slugify the provided tag string, only
+/// use the content you provide.
 public struct TagCapsule: View {
 
     /// Create a tag capsule.
@@ -35,17 +33,22 @@ public struct TagCapsule: View {
         Text(tag)
             .padding(style.padding)
             .foregroundColor(style.foregroundColor)
-            
             .materialCapsuleBackground(with: style.backgroundMaterial)
             .background(Capsule().fill(style.backgroundColor))
-            .padding(style.borderWidth)
+            .padding(style.border.width)
             .background(Capsule()
                 .strokeBorder(
-                    style.borderColor,
-                    lineWidth: style.borderWidth
+                    style.border.color,
+                    lineWidth: style.border.width
                 )
             )
             .compositingGroup()
+            .shadow(
+                color: style.shadow.color,
+                radius: style.shadow.radius,
+                x: style.shadow.offsetX,
+                y: style.shadow.offsetY
+            )
             
     }
 }
@@ -70,32 +73,50 @@ private extension View {
         LinearGradient(colors: [.blue, .red], startPoint: .top, endPoint: .bottom)
             .ignoresSafeArea()
         
-        VStack {
-            HStack {
-                TagCapsule("standard-tag")
-                TagCapsule("standard-selected-tag")
-                    .font(.body.bold())
-                    .tagCapsuleStyle(.standardSelected)
+        ScrollView {
+            VStack {
+                HStack {
+                    TagCapsule("standard")
+                        .tagCapsuleStyle(.standard)
+                    TagCapsule("standard-selected")
+                        .tagCapsuleStyle(.standardSelected)
+                }
+                HStack {
+                    TagCapsule("spider-man")
+                        .tagCapsuleStyle(.spiderman)
+                    TagCapsule("spider-man-selected")
+                        .tagCapsuleStyle(.spidermanSelected)
+                }
             }
-            TagCapsule("spider-man")
-                .tagCapsuleStyle(.init(
-                    foregroundColor: .black,
-                    backgroundColor: .red,
-                    borderColor: .blue,
-                    borderWidth: 4,
-                    padding: .init(
-                        top: 10,
-                        leading: 20,
-                        bottom: 12,
-                        trailing: 20
-                    )
-                ))
-                .shadow(radius: 0, y: 2)
-            TagCapsule("material")
-                .tagCapsuleStyle(.init(
-                    backgroundColor: .clear,
-                    backgroundMaterial: .thick
-                ))
+            .padding(.top, 250)
         }
+    }
+}
+
+private extension TagCapsuleStyle {
+    
+    static var spiderman: Self {
+        .init(
+            foregroundColor: .black,
+            backgroundColor: .red,
+            backgroundMaterial: .thin,
+            border: .init(
+                color: .blue,
+                width: 4
+            ),
+            padding: .init(
+                top: 10,
+                leading: 20,
+                bottom: 12,
+                trailing: 20
+            )
+        )
+    }
+    
+    static var spidermanSelected: Self {
+        var style = Self.spiderman
+        style.backgroundMaterial = .none
+        style.shadow = .standardSelected
+        return style
     }
 }