+++ Sources/TagKit/Views/TagList.swift
@@ -87,6 +87,7 @@ public struct TagList<TagView: View>: View {
     }
 }
 
+@MainActor
 private extension TagList {
     var content: some View {
         var width = CGFloat.zero